Create ICD_QC
ICD code contains the code on QCing the diagnostic code data; this included converting of some of the ICD-10 codes present in the data into ICD-9; removal of the codes that could not be attributed to any ICD diagnosis, or are not diagnosable in women; removal of codes for procedures or internal administrative Meuhedet codes, due to focus on ICD-9 diagnosis in the study. No filtering is made with regard to frequency of the diagnoses at this point; after QC, there are 931 unique diagnoses in the dataset.